pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "python-dateutil>=2.9.0.post0",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.41",
    "pandas>=2.3.1",
    "mammoth>=1.9.1",
    "python-docx>=1.2.0",
    "openpyxl>=3.1.5",
    "docx>=0.2.4",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "reportlab>=4.4.2",
    "weasyprint>=65.1",
    "matplotlib>=3.10.3",
    "xlsxwriter>=3.2.5",
    "numpy>=2.3.1",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.0",
    "selenium>=4.34.2",
    "webdriver-manager>=4.0.2",
    "chromedriver-autoinstaller>=0.6.4",
    "schedule>=1.2.2",
    "apscheduler>=3.11.0",
]
